# Peek before attempting to deserialize `ActiveSupport::Messages::Metadata` with JSON