--- a/activesupport/lib/active_support/messages/metadata.rb
+++ b/activesupport/lib/active_support/messages/metadata.rb
@@ -37,7 +37,10 @@ module ActiveSupport
           end
 
           def extract_metadata(message)
-            data = JSON.decode(message) rescue nil
+            begin
+              data = JSON.decode(message) if message.start_with?("{")
+            rescue ::JSON::JSONError
+            end
 
             if data.is_a?(Hash) && data.key?("_rails")
               new(decode(data["_rails"]["message"]), data["_rails"]["exp"], data["_rails"]["pur"])