@rafaelfranca 2022-11-29
Can you explain better how you are using this? The `Metadata` class is not public API of the framework and should not be used directly in applications. It is hard to judge by looking in this PR if the use is legit or not.