@jonathanhefner 2022-11-29
I can reproduce with only `MessageVerifier`, i.e. without using `Messages::Metadata` directly.

Making `Metadata` aware of the serializer won't really help, though, because the message format is actually dependent on [whether there is a `expires_at`, `expires_in`, or `purpose`](https://github.com/rails/rails/blob/c6b227b95b631ead7ffd5c8984b16b0fbe5b3ca4/activesupport/lib/active_support/messages/metadata.rb#L19-L23).  For example:

```ruby
# frozen_string_literal: true
require "benchmark/ips"
require "active_support/all"

verifier = ActiveSupport::MessageVerifier.new("secret", serializer: Marshal)
data = "x" * 100
message_with_purpose = verifier.generate(data, purpose: "x")
message_without_purpose = verifier.generate(data)

puts "Active Support version: #{ActiveSupport.version}"

Benchmark.ips do |x|
  x.report("with purpose") do
    verifier.verify(message_with_purpose, purpose: "x")
  end

  x.report("without purpose") do
    verifier.verify(message_without_purpose)
  end

  x.compare!
end
```

```console
$ bin/rails r 46612.rb
Active Support version: 7.1.0.alpha
Warming up --------------------------------------
        with purpose     2.698k i/100ms
     without purpose     1.451k i/100ms
Calculating -------------------------------------
        with purpose     26.627k (± 5.9%) i/s -    134.900k in   5.094282s
     without purpose     14.241k (± 1.6%) i/s -     72.550k in   5.095661s

Comparison:
        with purpose:    26626.8 i/s
     without purpose:    14241.0 i/s - 1.87x  (± 0.00) slower
```

If we were to branch on the verifier's serializer, it would break `expires_at`, `expires_in`, and `purpose` for non-JSON serializers.

*However*, the root cause appears to be something we're doing with exceptions in the `development` environment.  I cannot reproduce the issue in `test` or `production`:

```console
$ RAILS_ENV=test bin/rails r 46612.rb
Active Support version: 7.1.0.alpha
Warming up --------------------------------------
        with purpose     2.727k i/100ms
     without purpose     2.665k i/100ms
Calculating -------------------------------------
        with purpose     26.725k (± 6.9%) i/s -    133.623k in   5.041607s
     without purpose     26.548k (± 2.7%) i/s -    133.250k in   5.022939s

Comparison:
        with purpose:    26725.3 i/s
     without purpose:    26548.0 i/s - same-ish: difference falls within error
```

```console
$ RAILS_ENV=production bin/rails r 46612.rb
Active Support version: 7.1.0.alpha
Warming up --------------------------------------
        with purpose     2.646k i/100ms
     without purpose     2.690k i/100ms
Calculating -------------------------------------
        with purpose     25.500k (± 6.2%) i/s -    127.008k in   5.011294s
     without purpose     26.650k (± 2.1%) i/s -    134.500k in   5.049181s

Comparison:
     without purpose:    26649.9 i/s
        with purpose:    25500.3 i/s - same-ish: difference falls within error
```

Also, it appears to be on `main` only:

```console
$ ruby 46612.rb
Active Support version: 7.0.4
Warming up --------------------------------------
        with purpose     2.751k i/100ms
     without purpose     2.834k i/100ms
Calculating -------------------------------------
        with purpose     27.248k (± 2.5%) i/s -    137.550k in   5.051521s
     without purpose     28.680k (± 1.3%) i/s -    144.534k in   5.040404s

Comparison:
     without purpose:    28680.2 i/s
        with purpose:    27248.5 i/s - 1.05x  (± 0.00) slower
```

@ianks What Rails version + environment was your measurement taken from?