@ianks 2022-12-17
This was from a production stackprofile.

> Can you explain better how you are using this? The Metadata class is not public API of the framework and should not be used directly in applications

This was from using `ActiveSupport::MessageVerifier` directly. The time spent initializing the exception seems curiously long, and maybe hard to reproduce with simple benchmarks (since this exception is initialized deep in the stack of a large codebase).