@ianks 2022-12-17
Also, we are verifying fairly large blobs (on the order of ~4mb). That may also affect the benchmarks @jonathanhefner