@jonathanhefner 2022-12-17
@ianks When I increase the message size to 4MB, the `without purpose` case (the slow case) actually becomes comparatively faster, because the `with purpose` case now has to generate a 4MB JSON string.

Could you modify the script I posted in https://github.com/rails/rails/pull/46612#issuecomment-1331290378 to reproduce your issue?  Is there something in your production environment that might be slowing down `Exception#initialize`?