@ianks 2022-12-17
> @ianks When I increase the message size to 4MB, the `without purpose` case (the slow case) actually becomes comparatively faster, because the `with purpose` case now has to generate a 4MB JSON string.

The benchmark you have is benchmarking the wrong thing, IMO. We don't ever actually ever hit the `with purpose` case...

The thing to benchmark is `verifier.verify(message_without_purpose)`, with my patch and without my patch.

> Is there something in your production environment that might be slowing down `Exception#initialize`?

In the production environment, `Exception#initialize` is being called way deeper in the stack which will directly impact `Exception#initialize` times (since it has a much deeper stack to traverse). The benchmark won't capture that.

In general, using `rescue` for control flow is not a good great idea (partly for reasons like this).