@jonathanhefner 2022-12-18
> > When I increase the message size to 4MB, the `without purpose` case (the slow case) actually becomes comparatively faster, because the `with purpose` case now has to generate a 4MB JSON string.
>
> The benchmark you have is benchmarking the wrong thing, IMO. We don't ever actually ever hit the `with purpose` case...

Sorry, I wrote "generate", but I meant "parse".

What I was trying to say was that your case — the `without purpose` case — is actually faster than the other case when verifying a 4MB message (at least, in the benchmark).

That said, in my original comment (https://github.com/rails/rails/pull/46612#issuecomment-1331290378), I mistook the equal(ish) `production` ips as a lack of replication.  I thought the performance costs of parsing JSON vs an exception were equally small compared to cryptographic verification.  It turns out they are equally *large* (for small messages).  The `with purpose` case is bottlenecked by parsing the JSON into Ruby objects, hence it becomes the slower case when messages are 4MB.  The `without purpose` case is bottlenecked by the exception, which your patch fixes.

I apologize for my confusion! :pray:

Since branching on the serializer would break `expires_at`, `expires_in`, and `purpose` for non-JSON serializers, I think your current approach is the way to go.